SVAServer/requirements.txt
fastapi==0.116.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.5
saver==0.0.2
tqdm==4.67.1
uvicorn==0.35.0
